+++ CHANGELOG.md
@@ -10,8 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 - The bastion's public key will be created from the private key submitted in the `Settings` page. ([#318][i318])
 - The unused temporary dir has been removed from the `Settings` page. ([#165][i165]) 
 
+### Fixed
+- Everytime a user logs in, the model is updated (some date) on audit. ([#163][i163])
+- PHPStan message about Yajra DataTables. ([#331][i331])
+
 [i318]: https://github.com/pacoorozco/ssham/issues/318
 [i165]: https://github.com/pacoorozco/ssham/issues/165
+[i163]: https://github.com/pacoorozco/ssham/issues/163
+[i331]: https://github.com/pacoorozco/ssham/issues/331
 
 ## 0.15.1 - 2022-05-27